fix: small fixes to ABDK findings
* use mapping instead of fixed-length array to store queued accounts in `AccountFactoryV3`
* ensure successor compatibility in `GearStakingV3`
* add collateral hints validation
* fix `creditAccounts(uint256 offset, uint256 limit)`
* `getLTRampParamsHash` returns `bytes32` instead of `uint256`
* tiny performance improvements
* add some missing events and comments

+++ contracts/interfaces/IBotListV3.sol
@@ -12,6 +12,9 @@ struct BotFunding {
     uint40 allowanceLU;
 }
 
+/// @notice Bot special status parameters
+/// @param forbidden Whether bot is forbidden in the credit manager
+/// @param specialPermissions Mask of multicall actions that bot can perform for all accounts in the credit manager
 struct BotSpecialStatus {
     bool forbidden;
     uint192 specialPermissions;
@@ -23,7 +26,7 @@ interface IBotListV3Events {
         address indexed creditManager,
         address indexed creditAccount,
         address indexed bot,
-        uint256 permissions,
+        uint192 permissions,
         uint72 totalFundingAllowance,
         uint72 weeklyFundingAllowance
     );
